Update length of month and year
Since leap years are actually every four years EXCEPT every 100 years but INCLUDING every four-hundred years, the actual number of days in a year is 365.2425 not 365.25. This means that the number of days in a month is (on average) 30.436875 days.
This means we have to update all the test data to conform to the correct calculations.